No double jump, updates jumping interaction
There is now an indicator that appears on the mirror, showing how much the jump is charged. The jump charging also decreases over time, and has a cap on the maximum jump charge.

Double jumping has also been removed, although the solution will need to be fixed, as the current solution simply only allows the player to jump at y level 0.